WIP (floats, collision etc.)

+++ Assets/Place.cs
@@ -19,7 +19,7 @@ namespace Assets
             Hospital, Bar, Home, Office, Shop, None
         }
 
-        public int MoneyIncrease = 0,
+        public float MoneyIncrease = 0,
             HappinessIncrease = 0,
             BoredomIncrease = 0;
